chore: rename test package and unexport functions
Rename the test package from set_test back to set and unexport functions that do
not need to be exported anymore for testing purposes.
